requirements.txt
Flask>=2.2,<3
Flask-MySQLdb==1.0.1
pymysql==1.0.3
python-dotenv==1.0.0
Flask-WTF==1.1.1
Flask-Limiter>=3,<4
Flask-Talisman==1.0.0
Werkzeug>=2.2,<2.3
requests>=2.26.0,<2.27.0
gunicorn>=20.1.0
pathspec==0.9.0
redis